@@ -29,12 +29,20 @@
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
 		"prettier-plugin-tailwindcss": "^0.5.1",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0-next.164",
 		"svelte-check": "^3.6.0",
 		"tailwindcss": "^3.4.1",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
 		"vite": "^5.0.3"
 	},
-	"type": "module"
+	"type": "module",
+	"overrides": {
+		"flowbite-svelte": {
+			"svelte": "^5.0.0"
+		},
+		"flowbite-svelte-icons": {
+			"svelte": "^5.0.0"
+		}
+	}
 }
